+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* BraintreePayPal
+  * Add PayPal edit funding instrument flow (BETA)
+    * Add `BTPayPalEditRequest` for edit FI flow
+    * Add `BTPayPalClient.tokenize(_:completion:)` method that takes in a `BTPayPalEditRequest`
+    * **Note:** This feature is currently in beta and may change or be removed in future releases
+  
 ## 6.22.0 (2024-07-02)
 * BraintreeThreeDSecure
   * Add `customFields` param to `BTThreeDSecureRequest`